initial dre home page commit
