+++ dashboard-ui/src/components/AggregateResults/HomePages/MRE_homePage.jsx
@@ -1,9 +1,9 @@
 import { ScatterChart } from '@mui/x-charts';
 import React from 'react';
-import './aggregateResults.css';
+import '../aggregateResults.css';
 import CanvasJSReact from '@canvasjs/react-charts';
-import { getBoxWhiskerData, getMean, getMedian, getStandDev, getStandardError } from './statistics';
-import { isDefined, getChartData } from './DataFunctions';
+import { getBoxWhiskerData, getMean, getMedian, getStandDev, getStandardError } from '../statistics';
+import { isDefined, getChartData } from '../DataFunctions';
 
 
 const CanvasJSChart = CanvasJSReact.CanvasJSChart;
