+++ dashboard-ui/src/components/AggregateResults/aggregateResults.jsx
@@ -5,7 +5,7 @@ import { getAggregatedData, populateDataSet, isDefined, getGroupKey, formatCellD
 import * as FileSaver from 'file-saver';
 import XLSX from 'sheetjs-style';
 import './aggregateResults.css';
-import ProgramQuestions from './programQuestions';
+import ProgramQuestions from './HomePages/programQuestions';
 import { Modal } from "@mui/material";
 import { DefinitionTable } from './definitionTable';
 import CloseIcon from '@material-ui/icons/Close';
